Re-write conductor for clarity

+++ src/main/java/tk/jasoryeh/conductor/util/FileUtils.java
@@ -12,6 +12,10 @@ public class FileUtils {
     @SneakyThrows
     public static boolean delete(File f) {
         L.s(FILE_DELETE_LOG_PREFIX, "Delete: " + f.getAbsolutePath());
+        if (!f.exists()) {
+            L.s("File at " + f.getAbsolutePath() + " doesn't exist, we will assume it's already deleted.");
+            return true;
+        }
         if(Files.isSymbolicLink(f.toPath())) {
             Files.delete(f.toPath());
         }
@@ -24,21 +28,32 @@ public class FileUtils {
         return !f.exists();
     }
 
-    public static void deleteFile(File f) {
+    public static boolean deleteFile(File f) {
         L.s(FILE_DELETE_LOG_PREFIX, "Deleting file: " + f.getAbsolutePath());
+        if (!f.exists()) {
+            L.s("File at " + f.getAbsolutePath() + " doesn't exist, we will assume it's already deleted.");
+            return true;
+        }
         try {
             if(!f.delete()) {
                 L.s(FILE_DELETE_LOG_PREFIX, "Couldn't delete file: " + f.getAbsolutePath());
+                return false;
             }
         } catch(Exception e) {
-            L.s(FILE_DELETE_LOG_PREFIX, "Couldn't delete file: " + f.getAbsolutePath());
+            L.s(FILE_DELETE_LOG_PREFIX, "Failed to delete file: " + f.getAbsolutePath() + "; " + e.getMessage());
             e.printStackTrace();
+            return false;
         }
         L.s(FILE_DELETE_LOG_PREFIX, "File deleted: " + f.getAbsolutePath());
+        return !f.exists();
     }
 
-    public static void deleteFolder(File f) {
+    public static boolean deleteFolder(File f) {
         L.s(FILE_DELETE_LOG_PREFIX, "Deleting directory: " + f.getAbsolutePath());
+        if (!f.exists()) {
+            L.s("File at " + f.getAbsolutePath() + " doesn't exist, we will assume it's already deleted.");
+            return true;
+        }
         for (File file : f.listFiles()) {
             L.s(FILE_DELETE_LOG_PREFIX, "Emptying directory file: " + file.getAbsolutePath());
             if(file.isDirectory()) {
@@ -49,6 +64,7 @@ public class FileUtils {
         }
         deleteFile(f);
         L.s(FILE_DELETE_LOG_PREFIX, "Folder deleted: " + f.getAbsolutePath());
+        return !f.exists();
     }
 
 }
